requirements.txt
# Say What? - Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Video Extraction
yt-dlp==2024.1.31

# Transcription
openai-whisper==20231117
youtube-transcript-api==0.6.2

# AI Providers
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.4.0

# Vector Store & Embeddings
chromadb==0.4.22
tiktoken==0.6.0

# Settings & Validation
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
langdetect==1.0.9
python-slugify==8.0.2

# Session Management
itsdangerous==2.1.2
